build: Make it so that the requirements are 3.8 and 3.12 compatible.
The backports.zoneinfo package is not needed after python 3.8 so if we
run `make upgrade` with python 3.12 it doesn't get added but it is still
needed for Python 3.8 testing.  So we explicitly add it to `base.in`
with an environment marker so that it gets added for 3.8 but not 3.12 at
install time and we have requirements that are forward compatible.

We put the restriction in test.in because putting it in base.in causes
issues with how we pull that requirement via the `load_requirements`
function in setup.py